Move query into folder, remove delete arg
