+++ parseArguments.go
@@ -1,25 +1,21 @@
 package main
 
-import (
-	"flag"
-)
-
 /// Field Names Configuration: sets up how to parse task fields
 
-var FieldNamesConfig FieldNames
+// var FieldNamesConfig FieldNames
 
-func setupGlobalFieldNameFlags(flagsets []*flag.FlagSet) func() {
-	names := FieldNames{}
-
-	for _, fs := range flagsets {
-		fs.StringVar(&names.TitleField, "title", "Name", "Name of title [text] field")
-		fs.StringVar(&names.DoneField, "done", "Done", "Name of done [checkbox] field")
-		fs.StringVar(&names.DateField, "date", "Due", "Name of the date [date] field")
-		fs.StringVar(&names.CategoryField, "category", "Class", "Name of the category [select] field")
-		fs.StringVar(&names.TagField, "tag", "Tags", "Name of the tag [multi-select] field")
-	}
-
-	return func() {
-		FieldNamesConfig = names
-	}
-}
+// func setupGlobalFieldNameFlags(flagsets []*flag.FlagSet) func() {
+// names := FieldNames{}
+//
+// for _, fs := range flagsets {
+// fs.StringVar(&names.TitleField, "title", "Name", "Name of title [text] field")
+// fs.StringVar(&names.DoneField, "done", "Done", "Name of done [checkbox] field")
+// fs.StringVar(&names.DateField, "date", "Due", "Name of the date [date] field")
+// fs.StringVar(&names.CategoryField, "category", "Class", "Name of the category [select] field")
+// fs.StringVar(&names.TagField, "tag", "Tags", "Name of the tag [multi-select] field")
+// }
+//
+// return func() {
+// FieldNamesConfig = names
+// }
+// }
